+++ sulong/CHANGELOG.md
@@ -4,6 +4,8 @@ Changes:
 
 New features:
 
+* Fail when a function is called over interop with fewer arguments than it expects.
+
 * Support `InteropLibrary.isIdentical` and related messages.
 
 * Added the `--print-toolchain-api-identifier`,  `--print-toolchain-api-tool <tool-name>`